Read Me First
CVO is intended to be used for educational purposes regarding the basics of how video game hacking works. CVO uses the Windows API functions including GetModuleBaseAddress(), ReadProcessMemory() and WriteProcessMemory(). CVO is not intended to be used in public matches and you agree to being responsible for your own actions including receiving a possible game ban if you were to cheat with any anti-cheat enabled.